# docs/inji-certify-openapi.yaml
openapi: 3.0.1
info:
  version: v0
  title: OpenAPI definition
servers:
  - url: 'http://localhost:8090/v1/certify'
    description: Generated server url
paths:
  '/credential-configurations/{configurationId}':
    get:
      tags:
        - Credential Configuration
      summary: Fetch credential configuration
      description: Retrieves a credential configuration by its unique ID. Returns an error if the configuration does not exist or is not active.
      operationId: getCredentialConfigurationById
      parameters:
        - name: configurationId
          description: Unique key identifier for the credential configuration.
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credential configuration found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialConfigurationDTO'
              examples:
                Example 1:
                  value:
                    vcTemplate: eyJjcmVkZW50aWFsU3ViamVjdCI6eyJSZWdpc3RyYXRpb25EYXRlIjoiJHtyZWdpc3RyYXRpb25fZGF0ZX0iLCJDb25zZXJ2YXRpb25Vbml0UHJvY2Vzc2luZ0RhdGUiOiIke2NvbnNlcnZhdGlvbl91bml0X3Byb2Nlc3NpbmdfZGF0ZX0iLCJMYXRpdHVkZSI6IiR7bGF0aXR1ZGV9IiwiQ29uc29saWRhdGVkUnVyYWxBcmVhIjoiJHtjb25zb2xpZGF0ZWRfcnVyYWxfYXJlYX0iLCJMZWdhbFJlc2VydmVMb2NhdGlvbiI6IiR7bG9jYXRpb25fb2ZfbGVnYWxfcmVzZXJ2ZX0iLCJSZWdpc3RlcmVkTGVnYWxSZXNlcnZlQXJlYSI6IiR7cmVnaXN0ZXJlZF9sZWdhbF9yZXNlcnZlX2FyZWF9IiwiUmVzdHJpY3RlZFVzZUFyZWFUb0JlUmVzdG9yZWQiOiIke3Jlc3RyaWN0ZWRfdXNlX2FyZWFfdG9fcmVjb21wb3NlfSIsIlRvdGFsRGVjbGFyZWRMZWdhbFJlc2VydmUiOiIke3RvdGFsX2xlZ2FsX3Jlc2VydmVfZGVjbGFyZWR9IiwiUGVybWFuZW50UHJlc2VydmF0aW9uQXJlYXMiOiIke3Blcm1hbmVudF9wcmVzZXJ2YXRpb25fYXJlYXNfYXBwfSIsIkFkbWluaXN0cmF0aXZlRWFzZW1lbnRBcmVhIjoiJHthZG1pbmlzdHJhdGl2ZV9lYXNlbWVudF9hcmVhfSIsIlJlc3RyaWN0ZWRVc2VBcmVhcyI6IiR7cmVzdHJpY3RlZF91c2VfYXJlYXN9IiwiUFJBQ29uZGl0aW9uIjoiJHtwcmFfY29uZGl0aW9ufSIsImlkIjoiJHtfaG9sZGVySWR9IiwiTGVnYWxSZXNlcnZlRGVmaWNpdE9yRXhjZXNzIjoiJHtsZWdhbF9yZXNlcnZlX2RlZmljaXRfZXhjZXNzfSIsIkxlZ2FsUmVzZXJ2ZUFyZWFUb0JlUmVzdG9yZWQiOiIke2xlZ2FsX3Jlc2VydmVfYXJlYV90b19yZWNvbXBvc2V9IiwiRW1iYXJnZ2VkQXJlYU92ZXJsYXBQZXJjZW50YWdlIjoiJHtlbWJhcmdvZWRfYXJlYV9vdmVybGFwX3BlcmNlbnRhZ2V9IiwiRW1iYXJnZ2VkQXJlYURlc2NyaXB0aW9uIjoiJHtlbWJhcmdvZWRfYXJlYV9kZXNjcmlwdGlvbn0iLCJHZW9yZWZlcmVuY2VkTGVnYWxSZXNlcnZlQXJlYSI6IiR7Z2VvcmVmZXJlbmNlZF9sZWdhbF9yZXNlcnZlX2FyZWF9IiwiQVBQSW5Db25zb2xpZGF0ZWRSdXJhbEFyZWEiOiIke2FwcF9pbl9jb25zb2xpZGF0ZWRfcnVyYWxfYXJlYX0iLCJFbWJhcmdnZWRBcmVhUHJvY2Vzc2luZ0RhdGUiOiIke2VtYmFyZ29lZF9hcmVhX3Byb2Nlc3NpbmdfZGF0ZX0iLCJDb25zZXJ2YXRpb25Vbml0RGVzY3JpcHRpb24iOiIke2NvbnNlcnZhdGlvbl91bml0X2Rlc2NyaXB0aW9ufSIsIk11bmljaXBhbGl0eSI6IiR7bXVuaWNpcGFsaXR5fSIsIkVtYmFyZ2dlZEFyZWFPdmVybGFwIjoiJHtlbWJhcmdvZWRfYXJlYV9vdmVybGFwfSIsIkFwcHJvdmVkVW5yZWdpc3RlcmVkTGVnYWxSZXNlcnZlQXJlYSI6IiR7YXBwcm92ZWRfbGVnYWxfcmVzZXJ2ZV9hcmVhX25vdF9yZWdpc3RlcmVkfSIsIlJlZ2lzdHJhdGlvblN0YXR1cyI6IiR7cmVnaXN0cmF0aW9uX3N0YXR1c30iLCJOdW1iZXJPZkNBUiI6IiR7Y2FyX3JlZ2lzdHJhdGlvbl9udW1iZXJ9IiwiUnVyYWxQcm9wZXJ0eUFyZWEiOiIke3J1cmFsX3Byb3BlcnR5X2FyZWF9IiwiTG9uZ2l0dWRlIjoiJHtsb25naXR1ZGV9IiwiQ29uc2VydmF0aW9uVW5pdE92ZXJsYXBBcmVhIjoiJHtjb25zZXJ2YXRpb25fdW5pdF9vdmVybGFwX2FyZWF9IiwiRXh0ZXJuYWxDb25kaXRpb24iOiIke2V4dGVybmFsX2NvbmRpdGlvbn0iLCJMYXN0TW9kaWZpY2F0aW9uRGF0ZSI6IiR7ZGF0ZV9vZl9sYXN0X2FtZW5kbWVudH0iLCJDb25zZXJ2YXRpb25Vbml0T3ZlcmxhcFBlcmNlbnRhZ2UiOiIke2NvbnNlcnZhdGlvbl91bml0X292ZXJsYXBfcGVyY2VudGFnZX0iLCJOYXRpdmVWZWdldGF0aW9uUmVtbmFudEFyZWEiOiIke25hdGl2ZV92ZWdldGF0aW9uX3JlbW5hbnRfYXJlYX0iLCJBUFBJbk5hdGl2ZVZlZ2V0YXRpb25SZW1uYW50QXJlYSI6IiR7YXBwX2luX25hdGl2ZV92ZWdldGF0aW9uX3JlbW5hbnRfYXJlYX0iLCJQZXJtYW5lbnRQcmVzZXJ2YXRpb25BcmVhc1RvQmVSZXN0b3JlZCI6IiR7cGVybWFuZW50X3ByZXNlcnZhdGlvbl9hcmVhc190b19yZWNvbXBvc2V9IiwiR3JvdW5kQ292ZXJhZ2UiOiIke2xhbmRfY292ZXJ9IiwiUHJvcG9zZWRMZWdhbFJlc2VydmVBcmVhIjoiJHtwcm9wb3NlZF9sZWdhbF9yZXNlcnZlX2FyZWF9In0sInZhbGlkVW50aWwiOiIke3ZhbGlkVW50aWx9IiwidmFsaWRGcm9tIjoiJHt2YWxpZEZyb219IiwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsImxhbmRzdGF0ZW1lbnRfZGF0YUludGVncml0eSJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvbnMvY3JlZGVudGlhbHMvdjIiLCJodHRwczovL21vc2lwLmdpdGh1Yi5pby9pbmppLWNvbmZpZy9jb250ZXh0cy9sYW5kcmVnaXN0cnktc3RhdGVtZW50LWRhdGFJbnRlZ3JpdHktY29udGV4dC1hdXRvbWF0aW9uLmpzb24iXSwiaXNzdWVyIjoiJHtfaXNzdWVyfSJ9
                    keyManagerRefId: ED25519_SIGN
                    credentialSubjectDefinition:
                      APPInConsolidatedRuralArea:
                        display:
                          - name: APP in Consolidated Rural Area
                            locale: en
                      PermanentPreservationAreas:
                        display:
                          - name: Permanent Preservation Areas
                            locale: en
                    displayOrder:
                      - NumberOfCAR
                      - RegistrationDate
                      - LastModificationDate
                      - RuralPropertyArea
                      - Latitude
                      - Longitude
                      - Municipality
                      - ExternalCondition
                      - RegistrationStatus
                      - PRACondition
                      - GroundCoverage
                      - NativeVegetationRemnantArea
                      - ConsolidatedRuralArea
                      - AdministrativeEasementArea
                      - LegalReserveLocation
                      - RegisteredLegalReserveArea
                      - GeoreferencedLegalReserveArea
                      - ApprovedUnregisteredLegalReserveArea
                      - ProposedLegalReserveArea
                      - TotalDeclaredLegalReserve
                      - PermanentPreservationAreas
                      - APPInConsolidatedRuralArea
                      - APPInNativeVegetationRemnantArea
                      - RestrictedUseAreas
                      - LegalReserveDeficitOrExcess
                      - LegalReserveAreaToBeRestored
                      - PermanentPreservationAreasToBeRestored
                    didUrl: 'did:web:www.example.com:inji-config:collab:landregistry'
                    contextURLs:
                      - 'https://www.w3.org/ns/credentials/v2'
                      - 'https://www.example.com/inji-config/contexts/landregistry-statement-dataIntegrity-context-automation.json'
                    signatureCryptoSuite: eddsa-rdfc-2022
                    credentialConfigKeyId: landstatement_dataIntegrity_2.0
                    metaDataDisplay:
                      - background_image:
                          uri: 'https://www.example.com/inji-config/logos/agro-vertias-logo.png'
                        background_color: '#ebfaff'
                        name: land statement for data integrity
                        logo:
                          alt_text: land statement Logo
                          url: 'https://www.example.com/inji-config/logos/agro-vertias-logo.png'
                        text_color: '#000000'
                        locale: en
                    credentialFormat: ldp_vc
                    scope: land_statement_vc_ldp
                    keyManagerAppId: CERTIFY_VC_SIGN_ED25519
                    credentialTypes:
                      - VerifiableCredential
                      - landstatement_dataIntegrity
                    signatureAlgo: EdDSA
                    qrSettings:
                      - Full Name: "${fullName}"
                        Phone Number: "${mobileNumber}"
                        Date Of Birth: "${dateOfBirth}"
                    qrSignatureAlgo: EdDSA
        '404':
          description: Credential configuration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseWrapper'
              examples:
                Example 1:
                  value:
                    responseTime: '2025-08-07T13:02:15.071Z'
                    response: null
                    errors:
                      - errorCode: invalid_request
                        errorMessage: 'Configuration not found with the provided id: MockVerifiableCredential_SD_JWT2'
      servers:
        - url: 'http://localhost:8090/v1/certify'
          description: Generated server url
    put:
      tags:
        - Credential Configuration
      summary: Update credential configuration
      description: Updates an existing credential configuration. Performs similar validations as the creation endpoint.
      operationId: updateCredentialConfiguration
      parameters:
        - name: configurationId
          description: Unique key identifier for the credential configuration.
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialConfigurationDTO'
            examples:
              Example 1:
                value:
                  vcTemplate: eyJjcmVkZW50aWFsU3ViamVjdCI6eyJSZWdpc3RyYXRpb25EYXRlIjoiJHtyZWdpc3RyYXRpb25fZGF0ZX0iLCJDb25zZXJ2YXRpb25Vbml0UHJvY2Vzc2luZ0RhdGUiOiIke2NvbnNlcnZhdGlvbl91bml0X3Byb2Nlc3NpbmdfZGF0ZX0iLCJMYXRpdHVkZSI6IiR7bGF0aXR1ZGV9IiwiQ29uc29saWRhdGVkUnVyYWxBcmVhIjoiJHtjb25zb2xpZGF0ZWRfcnVyYWxfYXJlYX0iLCJMZWdhbFJlc2VydmVMb2NhdGlvbiI6IiR7bG9jYXRpb25fb2ZfbGVnYWxfcmVzZXJ2ZX0iLCJSZWdpc3RlcmVkTGVnYWxSZXNlcnZlQXJlYSI6IiR7cmVnaXN0ZXJlZF9sZWdhbF9yZXNlcnZlX2FyZWF9IiwiUmVzdHJpY3RlZFVzZUFyZWFUb0JlUmVzdG9yZWQiOiIke3Jlc3RyaWN0ZWRfdXNlX2FyZWFfdG9fcmVjb21wb3NlfSIsIlRvdGFsRGVjbGFyZWRMZWdhbFJlc2VydmUiOiIke3RvdGFsX2xlZ2FsX3Jlc2VydmVfZGVjbGFyZWR9IiwiUGVybWFuZW50UHJlc2VydmF0aW9uQXJlYXMiOiIke3Blcm1hbmVudF9wcmVzZXJ2YXRpb25fYXJlYXNfYXBwfSIsIkFkbWluaXN0cmF0aXZlRWFzZW1lbnRBcmVhIjoiJHthZG1pbmlzdHJhdGl2ZV9lYXNlbWVudF9hcmVhfSIsIlJlc3RyaWN0ZWRVc2VBcmVhcyI6IiR7cmVzdHJpY3RlZF91c2VfYXJlYXN9IiwiUFJBQ29uZGl0aW9uIjoiJHtwcmFfY29uZGl0aW9ufSIsImlkIjoiJHtfaG9sZGVySWR9IiwiTGVnYWxSZXNlcnZlRGVmaWNpdE9yRXhjZXNzIjoiJHtsZWdhbF9yZXNlcnZlX2RlZmljaXRfZXhjZXNzfSIsIkxlZ2FsUmVzZXJ2ZUFyZWFUb0JlUmVzdG9yZWQiOiIke2xlZ2FsX3Jlc2VydmVfYXJlYV90b19yZWNvbXBvc2V9IiwiRW1iYXJnZ2VkQXJlYU92ZXJsYXBQZXJjZW50YWdlIjoiJHtlbWJhcmdvZWRfYXJlYV9vdmVybGFwX3BlcmNlbnRhZ2V9IiwiRW1iYXJnZ2VkQXJlYURlc2NyaXB0aW9uIjoiJHtlbWJhcmdvZWRfYXJlYV9kZXNjcmlwdGlvbn0iLCJHZW9yZWZlcmVuY2VkTGVnYWxSZXNlcnZlQXJlYSI6IiR7Z2VvcmVmZXJlbmNlZF9sZWdhbF9yZXNlcnZlX2FyZWF9IiwiQVBQSW5Db25zb2xpZGF0ZWRSdXJhbEFyZWEiOiIke2FwcF9pbl9jb25zb2xpZGF0ZWRfcnVyYWxfYXJlYX0iLCJFbWJhcmdnZWRBcmVhUHJvY2Vzc2luZ0RhdGUiOiIke2VtYmFyZ29lZF9hcmVhX3Byb2Nlc3NpbmdfZGF0ZX0iLCJDb25zZXJ2YXRpb25Vbml0RGVzY3JpcHRpb24iOiIke2NvbnNlcnZhdGlvbl91bml0X2Rlc2NyaXB0aW9ufSIsIk11bmljaXBhbGl0eSI6IiR7bXVuaWNpcGFsaXR5fSIsIkVtYmFyZ2dlZEFyZWFPdmVybGFwIjoiJHtlbWJhcmdvZWRfYXJlYV9vdmVybGFwfSIsIkFwcHJvdmVkVW5yZWdpc3RlcmVkTGVnYWxSZXNlcnZlQXJlYSI6IiR7YXBwcm92ZWRfbGVnYWxfcmVzZXJ2ZV9hcmVhX25vdF9yZWdpc3RlcmVkfSIsIlJlZ2lzdHJhdGlvblN0YXR1cyI6IiR7cmVnaXN0cmF0aW9uX3N0YXR1c30iLCJOdW1iZXJPZkNBUiI6IiR7Y2FyX3JlZ2lzdHJhdGlvbl9udW1iZXJ9IiwiUnVyYWxQcm9wZXJ0eUFyZWEiOiIke3J1cmFsX3Byb3BlcnR5X2FyZWF9IiwiTG9uZ2l0dWRlIjoiJHtsb25naXR1ZGV9IiwiQ29uc2VydmF0aW9uVW5pdE92ZXJsYXBBcmVhIjoiJHtjb25zZXJ2YXRpb25fdW5pdF9vdmVybGFwX2FyZWF9IiwiRXh0ZXJuYWxDb25kaXRpb24iOiIke2V4dGVybmFsX2NvbmRpdGlvbn0iLCJMYXN0TW9kaWZpY2F0aW9uRGF0ZSI6IiR7ZGF0ZV9vZl9sYXN0X2FtZW5kbWVudH0iLCJDb25zZXJ2YXRpb25Vbml0T3ZlcmxhcFBlcmNlbnRhZ2UiOiIke2NvbnNlcnZhdGlvbl91bml0X292ZXJsYXBfcGVyY2VudGFnZX0iLCJOYXRpdmVWZWdldGF0aW9uUmVtbmFudEFyZWEiOiIke25hdGl2ZV92ZWdldGF0aW9uX3JlbW5hbnRfYXJlYX0iLCJBUFBJbk5hdGl2ZVZlZ2V0YXRpb25SZW1uYW50QXJlYSI6IiR7YXBwX2luX25hdGl2ZV92ZWdldGF0aW9uX3JlbW5hbnRfYXJlYX0iLCJQZXJtYW5lbnRQcmVzZXJ2YXRpb25BcmVhc1RvQmVSZXN0b3JlZCI6IiR7cGVybWFuZW50X3ByZXNlcnZhdGlvbl9hcmVhc190b19yZWNvbXBvc2V9IiwiR3JvdW5kQ292ZXJhZ2UiOiIke2xhbmRfY292ZXJ9IiwiUHJvcG9zZWRMZWdhbFJlc2VydmVBcmVhIjoiJHtwcm9wb3NlZF9sZWdhbF9yZXNlcnZlX2FyZWF9In0sInZhbGlkVW50aWwiOiIke3ZhbGlkVW50aWx9IiwidmFsaWRGcm9tIjoiJHt2YWxpZEZyb219IiwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsImxhbmRzdGF0ZW1lbnRfZGF0YUludGVncml0eSJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvbnMvY3JlZGVudGlhbHMvdjIiLCJodHRwczovL21vc2lwLmdpdGh1Yi5pby9pbmppLWNvbmZpZy9jb250ZXh0cy9sYW5kcmVnaXN0cnktc3RhdGVtZW50LWRhdGFJbnRlZ3JpdHktY29udGV4dC1hdXRvbWF0aW9uLmpzb24iXSwiaXNzdWVyIjoiJHtfaXNzdWVyfSJ9
                  keyManagerRefId: ED25519_SIGN
                  credentialSubjectDefinition:
                    APPInConsolidatedRuralArea:
                      display:
                        - name: APP in Consolidated Rural Area
                          locale: en
                    PermanentPreservationAreas:
                      display:
                        - name: Permanent Preservation Areas
                          locale: en
                  displayOrder:
                    - NumberOfCAR
                    - RegistrationDate
                    - LastModificationDate
                    - RuralPropertyArea
                    - Latitude
                    - Longitude
                    - Municipality
                    - ExternalCondition
                    - RegistrationStatus
                    - PRACondition
                    - GroundCoverage
                    - NativeVegetationRemnantArea
                    - ConsolidatedRuralArea
                    - AdministrativeEasementArea
                    - LegalReserveLocation
                    - RegisteredLegalReserveArea
                    - GeoreferencedLegalReserveArea
                    - ApprovedUnregisteredLegalReserveArea
                    - ProposedLegalReserveArea
                    - TotalDeclaredLegalReserve
                    - PermanentPreservationAreas
                    - APPInConsolidatedRuralArea
                    - APPInNativeVegetationRemnantArea
                    - RestrictedUseAreas
                    - LegalReserveDeficitOrExcess
                    - LegalReserveAreaToBeRestored
                    - PermanentPreservationAreasToBeRestored
                  didUrl: 'did:web:www.example.com:inji-config:collab:landregistry'
                  contextURLs:
                    - 'https://www.w3.org/ns/credentials/v2'
                    - 'https://www.example.com/inji-config/contexts/landregistry-statement-dataIntegrity-context-automation.json'
                  signatureCryptoSuite: eddsa-rdfc-2022
                  credentialConfigKeyId: landstatement_dataIntegrity_2.0
                  metaDataDisplay:
                    - background_image:
                        uri: 'https://www.example.com/inji-config/logos/agro-vertias-logo.png'
                      background_color: '#ebfaff'
                      name: land statement for data integrity
                      logo:
                        alt_text: land statement Logo
                        url: 'https://www.example.com/inji-config/logos/agro-vertias-logo.png'
                      text_color: '#000000'
                      locale: en
                  credentialFormat: ldp_vc
                  scope: land_statement_vc_ldp
                  keyManagerAppId: CERTIFY_VC_SIGN_ED25519
                  credentialTypes:
                    - VerifiableCredential
                    - landstatement_dataIntegrity
                  signatureAlgo: EdDSA
                  qrSettings:
                    - Full Name: "${fullName}"
                      Phone Number: "${mobileNumber}"
                      Date Of Birth: "${dateOfBirth}"
                  qrSignatureAlgo: EdDSA
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialConfigResponse'
              examples:
                Example 1:
                  value:
                    vcTemplate: eyJjcmVkZW50aWFsU3ViamVjdCI6eyJSZWdpc3RyYXRpb25EYXRlIjoiJHtyZWdpc3RyYXRpb25fZGF0ZX0iLCJDb25zZXJ2YXRpb25Vbml0UHJvY2Vzc2luZ0RhdGUiOiIke2NvbnNlcnZhdGlvbl91bml0X3Byb2Nlc3NpbmdfZGF0ZX0iLCJMYXRpdHVkZSI6IiR7bGF0aXR1ZGV9IiwiQ29uc29saWRhdGVkUnVyYWxBcmVhIjoiJHtjb25zb2xpZGF0ZWRfcnVyYWxfYXJlYX0iLCJMZWdhbFJlc2VydmVMb2NhdGlvbiI6IiR7bG9jYXRpb25fb2ZfbGVnYWxfcmVzZXJ2ZX0iLCJSZWdpc3RlcmVkTGVnYWxSZXNlcnZlQXJlYSI6IiR7cmVnaXN0ZXJlZF9sZWdhbF9yZXNlcnZlX2FyZWF9IiwiUmVzdHJpY3RlZFVzZUFyZWFUb0JlUmVzdG9yZWQiOiIke3Jlc3RyaWN0ZWRfdXNlX2FyZWFfdG9fcmVjb21wb3NlfSIsIlRvdGFsRGVjbGFyZWRMZWdhbFJlc2VydmUiOiIke3RvdGFsX2xlZ2FsX3Jlc2VydmVfZGVjbGFyZWR9IiwiUGVybWFuZW50UHJlc2VydmF0aW9uQXJlYXMiOiIke3Blcm1hbmVudF9wcmVzZXJ2YXRpb25fYXJlYXNfYXBwfSIsIkFkbWluaXN0cmF0aXZlRWFzZW1lbnRBcmVhIjoiJHthZG1pbmlzdHJhdGl2ZV9lYXNlbWVudF9hcmVhfSIsIlJlc3RyaWN0ZWRVc2VBcmVhcyI6IiR7cmVzdHJpY3RlZF91c2VfYXJlYXN9IiwiUFJBQ29uZGl0aW9uIjoiJHtwcmFfY29uZGl0aW9ufSIsImlkIjoiJHtfaG9sZGVySWR9IiwiTGVnYWxSZXNlcnZlRGVmaWNpdE9yRXhjZXNzIjoiJHtsZWdhbF9yZXNlcnZlX2RlZmljaXRfZXhjZXNzfSIsIkxlZ2FsUmVzZXJ2ZUFyZWFUb0JlUmVzdG9yZWQiOiIke2xlZ2FsX3Jlc2VydmVfYXJlYV90b19yZWNvbXBvc2V9IiwiRW1iYXJnZ2VkQXJlYU92ZXJsYXBQZXJjZW50YWdlIjoiJHtlbWJhcmdvZWRfYXJlYV9vdmVybGFwX3BlcmNlbnRhZ2V9IiwiRW1iYXJnZ2VkQXJlYURlc2NyaXB0aW9uIjoiJHtlbWJhcmdvZWRfYXJlYV9kZXNjcmlwdGlvbn0iLCJHZW9yZWZlcmVuY2VkTGVnYWxSZXNlcnZlQXJlYSI6IiR7Z2VvcmVmZXJlbmNlZF9sZWdhbF9yZXNlcnZlX2FyZWF9IiwiQVBQSW5Db25zb2xpZGF0ZWRSdXJhbEFyZWEiOiIke2FwcF9pbl9jb25zb2xpZGF0ZWRfcnVyYWxfYXJlYX0iLCJFbWJhcmdnZWRBcmVhUHJvY2Vzc2luZ0RhdGUiOiIke2VtYmFyZ29lZF9hcmVhX3Byb2Nlc3NpbmdfZGF0ZX0iLCJDb25zZXJ2YXRpb25Vbml0RGVzY3JpcHRpb24iOiIke2NvbnNlcnZhdGlvbl91bml0X2Rlc2NyaXB0aW9ufSIsIk11bmljaXBhbGl0eSI6IiR7bXVuaWNpcGFsaXR5fSIsIkVtYmFyZ2dlZEFyZWFPdmVybGFwIjoiJHtlbWJhcmdvZWRfYXJlYV9vdmVybGFwfSIsIkFwcHJvdmVkVW5yZWdpc3RlcmVkTGVnYWxSZXNlcnZlQXJlYSI6IiR7YXBwcm92ZWRfbGVnYWxfcmVzZXJ2ZV9hcmVhX25vdF9yZWdpc3RlcmVkfSIsIlJlZ2lzdHJhdGlvblN0YXR1cyI6IiR7cmVnaXN0cmF0aW9uX3N0YXR1c30iLCJOdW1iZXJPZkNBUiI6IiR7Y2FyX3JlZ2lzdHJhdGlvbl9udW1iZXJ9IiwiUnVyYWxQcm9wZXJ0eUFyZWEiOiIke3J1cmFsX3Byb3BlcnR5X2FyZWF9IiwiTG9uZ2l0dWRlIjoiJHtsb25naXR1ZGV9IiwiQ29uc2VydmF0aW9uVW5pdE92ZXJsYXBBcmVhIjoiJHtjb25zZXJ2YXRpb25fdW5pdF9vdmVybGFwX2FyZWF9IiwiRXh0ZXJuYWxDb25kaXRpb24iOiIke2V4dGVybmFsX2NvbmRpdGlvbn0iLCJMYXN0TW9kaWZpY2F0aW9uRGF0ZSI6IiR7ZGF0ZV9vZl9sYXN0X2FtZW5kbWVudH0iLCJDb25zZXJ2YXRpb25Vbml0T3ZlcmxhcFBlcmNlbnRhZ2UiOiIke2NvbnNlcnZhdGlvbl91bml0X292ZXJsYXBfcGVyY2VudGFnZX0iLCJOYXRpdmVWZWdldGF0aW9uUmVtbmFudEFyZWEiOiIke25hdGl2ZV92ZWdldGF0aW9uX3JlbW5hbnRfYXJlYX0iLCJBUFBJbk5hdGl2ZVZlZ2V0YXRpb25SZW1uYW50QXJlYSI6IiR7YXBwX2luX25hdGl2ZV92ZWdldGF0aW9uX3JlbW5hbnRfYXJlYX0iLCJQZXJtYW5lbnRQcmVzZXJ2YXRpb25BcmVhc1RvQmVSZXN0b3JlZCI6IiR7cGVybWFuZW50X3ByZXNlcnZhdGlvbl9hcmVhc190b19yZWNvbXBvc2V9IiwiR3JvdW5kQ292ZXJhZ2UiOiIke2xhbmRfY292ZXJ9IiwiUHJvcG9zZWRMZWdhbFJlc2VydmVBcmVhIjoiJHtwcm9wb3NlZF9sZWdhbF9yZXNlcnZlX2FyZWF9In0sInZhbGlkVW50aWwiOiIke3ZhbGlkVW50aWx9IiwidmFsaWRGcm9tIjoiJHt2YWxpZEZyb219IiwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsImxhbmRzdGF0ZW1lbnRfZGF0YUludGVncml0eSJdLCJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvbnMvY3JlZGVudGlhbHMvdjIiLCJodHRwczovL21vc2lwLmdpdGh1Yi5pby9pbmppLWNvbmZpZy9jb250ZXh0cy9sYW5kcmVnaXN0cnktc3RhdGVtZW50LWRhdGFJbnRlZ3JpdHktY29udGV4dC1hdXRvbWF0aW9uLmpzb24iXSwiaXNzdWVyIjoiJHtfaXNzdWVyfSJ9
                    keyManagerRefId: ED25519_SIGN
                    credentialSubjectDefinition:
                      APPInConsolidatedRuralArea:
                        display:
                          - name: APP in Consolidated Rural Area
                            locale: en
                      PermanentPreservationAreas:
                        display:
                          - name: Permanent Preservation Areas
                            locale: en
                    displayOrder:
                      - NumberOfCAR
                      - RegistrationDate
                      - LastModificationDate
                      - RuralPropertyArea
                      - Latitude
                      - Longitude
                      - Municipality
                      - ExternalCondition
                      - RegistrationStatus
                      - PRACondition
                      - GroundCoverage
                      - NativeVegetationRemnantArea
                      - ConsolidatedRuralArea
                      - AdministrativeEasementArea
                      - LegalReserveLocation
                      - RegisteredLegalReserveArea
                      - GeoreferencedLegalReserveArea
                      - ApprovedUnregisteredLegalReserveArea
                      - ProposedLegalReserveArea
                      - TotalDeclaredLegalReserve
                      - PermanentPreservationAreas
                      - APPInConsolidatedRuralArea
                      - APPInNativeVegetationRemnantArea
                      - RestrictedUseAreas
                      - LegalReserveDeficitOrExcess
                      - LegalReserveAreaToBeRestored
                      - PermanentPreservationAreasToBeRestored
                    didUrl: 'did:web:www.example.com:inji-config:collab:landregistry'
                    contextURLs:
                      - 'https://www.w3.org/ns/credentials/v2'
                      - 'https://www.example.com/inji-config/contexts/landregistry-statement-dataIntegrity-context-automation.json'
                    signatureCryptoSuite: eddsa-rdfc-2022
                    credentialConfigKeyId: landstatement_dataIntegrity_2.0
                    metaDataDisplay:
                      - background_image:
                          uri: 'https://www.example.com/inji-config/logos/agro-vertias-logo.png'
                        background_color: '#ebfaff'
                        name: land statement for data integrity
                        logo:
                          alt_text: land statement Logo
                          url: 'https://www.example.com/inji-config/logos/agro-vertias-logo.png'
                        text_color: '#000000'
                        locale: en
                    credentialFormat: ldp_vc
                    scope: land_statement_vc_ldp
                    keyManagerAppId: CERTIFY_VC_SIGN_ED25519
                    credentialTypes:
                      - VerifiableCredential
                      - landstatement_dataIntegrity
                    signatureAlgo: EdDSA
      servers:
        - url: 'http://localhost:8090/v1/certify'
          description: Generated server url
    delete:
      tags:
        - Credential Configuration
      summary: Delete credential configuration
      description: Deletes a credential configuration by its ID. Returns an error if the configuration does not exist.
      operationId: deleteCredentialConfigurationById
      parameters:
        - name: configurationId
          description: Unique key identifier for the credential configuration
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
      servers:
        - url: 'http://localhost:8090/v1/certify'
          description: Generated server url
...
components:
  schemas:
    CredentialConfigurationDTO:
      type: object
      required:
        - credentialFormat
        - metaDataDisplay
        - scope
      properties:
        vcTemplate:
          type: string
          description: Base64 encoded template identifier for the Verifiable Credential. Mandatory if mosip.certify.plugin-mode is set to ‘DataProvider’ during creation.
        credentialConfigKeyId:
          type: string
          description: Unique key identifier for the credential configuration.
        contextURLs:
          type: array
          description: 'List of JSON-LD context URIs for the credential. It should be list of valid URI strings, will be mandatory for ldp_vc credential format.'
          items:
            type: string
        credentialTypes:
          type: array
          description: 'List of credential types (e.g., VerifiableCredential,FarmerCredential) as OpenID VCI specifications. Mandatory and applicable only for ldp_vc format.'
          items:
            type: string
        sdJwtVct:
          type: string
          description: Represents the Verifiable Credential Type (VCT) used in Selective Disclosure JWT (SD-JWT) credentials. This field specifies the credential type identifier required for vc+sd-jwt format.
        credentialFormat:
          description: 'Format of the credential. Valid values are ldp_vc, mso_mdoc, vc+sd-jwt.'
          type: string
          enum:
            - ldp_vc
            - mso_mdoc
            - vc+sd-jwt
        didUrl:
          type: string
          description: URL of the Decentralized Identifier (DID) document. Mandatory for DataProvider plugin mode.
        keyManagerAppId:
          type: string
          description: Application ID for the key manager used in signing. Required for DataProvider plugin mode.
        keyManagerRefId:
          type: string
          description: Reference ID for the key manager. Applicable for DataProvider plugin mode.
        signatureAlgo:
          description: |-
            Signature or proof algorithm used for credential signing. This field is required for certain credential formats, such as vc+sd-jwt and optionally for ldp_vc when the signatureCryptoSuite is one of DataIntegrityProof suite (e.g. ecdsa-rdfc-2019, ecdsa-jcs-2019, etc). It ensures compatibility between the cryptographic suite and the signing algorithm. Mapping of signatureCryptoSuite and signatureAlgo values for DataIntegrityProof suite scenario :

            signatureCryptoSuite |  signatureAlgo
            ---------|----------
             eddsa-rdfc-2022, eddsa-jcs-2022 | EdDSA 
             ecdsa-rdfc-2019, ecdsa-jcs-2019 | ES256K,ES256 

          enum:
            - EdDSA
            - ES256K
            - ES256
          type: string
        sdClaim:
          type: string
          description: Comma-separated list for selective disclosure claim configuration. Applicable for vc+sd-jwt credential type.
        metaDataDisplay:
          type: array
          description: 'To configure the display properties (e.g., logo, name, colors, and localization) for metadata associated with the Verifiable Credential type.'
          items:
            $ref: '#/components/schemas/MetaDataDisplayDTO'
        displayOrder:
          type: array
          description: 'List specifying the order of claims or fields in the credential that are defined in credentialSubjectDefinition field, to be used while showing the credential in UI.'
          items:
            type: string
        scope:
          type: string
          description: Scope value for the credential configuration. This value is matched with the scope value in Auth Token while processing the credential issuance request.
        pluginConfigurations:
          type: array
          description: 'Holds a list of plugin-specific configuration parameters as key-value pairs. Each entry in the list represents configuration settings required by external plugins or modules integrated with the credential issuance process. This allows dynamic extension and customization of credential behavior based on the needs of different plugins, such as data providers, signature providers, or custom validation logic.'
          items:
            type: object
            additionalProperties:
              type: string
        credentialSubjectDefinition:
          type: object
          description: 'Represents the subject of the Verifiable Credential, containing the claims and attributes about the entity (such as a person, organization, or device) to whom the credential is issued. Applicable for ldp_vc credential type.'
          additionalProperties:
            $ref: '#/components/schemas/CredentialSubjectParametersDTO'
        msoMdocClaims:
          description: Claims included in the credential. Applicable for mso_mdoc and vc+sd-jwt credential formats.
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              $ref: '#/components/schemas/ClaimsDisplayFieldsConfigDTO'
        sdJwtClaims:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ClaimsDisplayFieldsConfigDTO'
        doctype:
          type: string
          description: Document type identifier for the credential. Mandatory for mso_mdoc credential format.
        signatureCryptoSuite:
          description: 'Holds value for signature crypto suites supported for key algorithms like Ed25519Signature2020,EcdsaSecp256k1Signature2019, etc. It will also be applicable for data integrity proof crypto suites, for e.g. ecds-rdfc-2019, eddsa-rdfc-2022, etc. Mandatory for DataProvider plugin. The values should be only be those supported by certify.'
          type: string
          enum:
            - RsaSignature2018
            - Ed25519Signature2018
            - Ed25519Signature2020
            - EcdsaKoblitzSignature2016
            - EcdsaSecp256k1Signature2019
            - EcdsaSecp256r1Signature2019
            - ecdsa-rdfc-2019
            - ecdsa-jcs-2019
            - eddsa-rdfc-2022
            - eddsa-jcs-2022
        credentialStatusPurposes:
          type: string
          description: 'List of purposes for which credential status is maintained, comma-separated values are supported here. Currently, the value for this field will be driven by following config - `mosip.certify.data-provider-plugin.credential-status.supported-purposes`. Value sent in the request will be ignored. '
        qrSettings:
          type: array
          description: List of QR settings for credential configuration. Each item is a map of key-value pairs; values may be primitives or nested objects.
          items:
            type: object
            additionalProperties:
              type: string
        qrSignatureAlgo:
          type: string
          description: Signature algorithm used for QR code signing.
    CredentialConfigResponse:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
      required:
        - id
        - status
...
    ClaimsDisplayFieldsConfigDTO:
      title: ClaimsDisplayFieldsConfigDTO
      type: object
      properties:
        display:
          type: array
          items:
            $ref: '#/components/schemas/Display'
    CredentialSubjectParametersDTO:
      type: object
      properties:
        display:
          type: array
          items:
            $ref: '#/components/schemas/Display'
...
    Logo:
      type: object
      properties:
        url:
          type: string
          description: The URL of the logo image.
        alt_text:
          type: string
          description: 'Alternative text for the logo, used for accessibility.'
    BackgroundImage:
      type: object
      properties:
        uri:
          type: string
          description: The URI of the background image.
    MetaDataDisplayDTO:
      type: object
      properties:
        logo:
          $ref: '#/components/schemas/Logo'
          description: Holds information regarding the logo image.
        name:
          type: string
          description: 'The display name, typically used as a title or label.'
        locale:
          type: string
          description: 'The language or region code for localization (e.g., en-US).'
        text_color:
          type: string
          description: 'The color code for text, used for UI customization.'
        background_color:
          type: string
          description: The color for background when credential is displayed.
        background_image:
          $ref: '#/components/schemas/BackgroundImage'
    ExceptionResponseWrapper:
      type: object
      x-examples:
        Example 1:
          responseTime: '2025-08-07T10:04:35.037Z'
          response: null
          errors:
            - errorCode: invalid_request
              errorMessage: 'Configuration not found with the provided id: MockVerifiableCredential_SD_JWT2'
      properties:
        responseTime:
          type: string
          description: 'The timestamp indicating when the response was generated, in ISO 8601 format.'
        response:
          description: 'The main response data. It is null in this case, indicating no successful data was returned.'
          nullable: true
        errors:
          type: array
          description: |
            An array containing error details if the request failed.
          items:
            type: object
            properties:
              errorCode:
                type: string
                description: A code representing the type of error encountered.
              errorMessage:
                type: string
                description: A human-readable message describing the error.
...